requirements.txt
python=3.6
numpy
pandas
pyyaml
